V1.1.2

+++ src/main/java/fr/potrunks/gestiondepensebackend/business/impl/SpentBusiness.java
@@ -60,9 +60,11 @@ public class SpentBusiness implements SpentIBusiness {
     @Override
     public Boolean deleteSpent(Long idSpent) {
         log.info("Start to find spent by ID {}", idSpent);
-        SpentEntity spentEntity = spentRepository.findById(idSpent).get();
-        log.info("Start to delete spent by ID {}", idSpent);
-        spentRepository.delete(spentEntity);
+        SpentEntity spentEntity = spentRepository.getById(idSpent);
+        if (spentEntity != null) {
+            log.info("Start to delete spent by ID {}", idSpent);
+            spentRepository.delete(spentEntity);
+        }
         return true;
     }
